requirements.txt
google-generativeai
pyautogui
requests
python-dotenv
